Upgrade psalm 5.26 to 6.1
Preparation for TikTok
- privacypolicy and termsofservice introduced

latest vimeo/psalm version 6.1 introduced - 186 errors fixed

+++ src/Invoice/ProductImage/ProductImageService.php
@@ -53,7 +53,11 @@ final class ProductImageService
         $targetPath = $aliases->get('@public_product_images');
         $file_path = $targetPath . '/' . $model->getFile_name_new();
         // see vendor/yiisoft/files/src/FileHelper::unlink will delete the file
-        strpos(realpath($targetPath), realpath($file_path)) == 0 ? FileHelper::unlink($file_path) : '';
-        $this->repository->delete($model);
+        $realTargetPath = realpath($targetPath);
+        $realFilePath = realpath($file_path);
+        if (($realTargetPath <> false) && ($realFilePath <> false)) {
+            strpos($realTargetPath, $realFilePath) == 0 ? FileHelper::unlink($file_path) : '';
+            $this->repository->delete($model);
+        }    
     }
 }
